requirements.txt
pandas
scipy
plotly.express
streamlit
